ObjC: Fix quality scaling for injected encoders
We missed to implement quality scaling in the original CL
https://codereview.webrtc.org/2977213002/. This CL implements it.

Note that the ObjC interface for scalingSettings is slightly different from the C++
interface in that we require explicit QP thresholds to turn quality scaling on, i.e.
we don't provide default values. I think this is more modular as we want to move
codec specific knowledge out from the WebRTC core. I would like to update the
C++ webrtc::VideoEncoder interface to do the same in another CL.

BUG=webrtc:7924

Review-Url: https://codereview.webrtc.org/2991123002
Cr-Original-Commit-Position: refs/heads/master@{#19202}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 5dfac33dfdd430d60f3ecf4c59d15a62ec1f68ef