Make it possible for tests to set up trace event handlers.
As it is now, the first time a TRACE_EVENT... is called, the result from
the current handler is stored in a static const variable, and subsequent
calls will use that value regardless of changes to the handler.

This is a problem if a test wants to use another handler.

BUG=None

Review-Url: https://codereview.webrtc.org/3002663002
Cr-Original-Commit-Position: refs/heads/master@{#19382}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: a9732656221e0033ab93ecf01cf07f71817da1e8